Refactor tests
Due to a quirk in the Artifactory API, all Repositories come with a
`projectKey` value of "default". This value is not kept in state, and
results in a permanent diff on `pulumi update`.
To be able to run complete tests, this commit switches the tests to use
the User resource instead.